reorganize roadrunner into Autonomous package
