+++ FtcRobotController-master/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/messages/TankCommandMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.messages;
+package org.firstinspires.ftc.teamcode.Autonomous.messages;
 
 public final class TankCommandMessage {
     public long timestamp;
