+++ FtcRobotController-master/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/messages/TankLocalizerInputsMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.messages;
+package org.firstinspires.ftc.teamcode.Autonomous.messages;
 
 import com.acmerobotics.roadrunner.ftc.PositionVelocityPair;
